what is it for?
This simple calculator serves to reveal the actual total of people infected in your region or country. because it is very simple, the user inputs the data, makes a very simple calculation and returns with the actual amount of infected!

For those who are beginners like, I will explain the concern that was made.

n1 = input ('Infected:')
n2 = input ('Cured:')
n3 = input ('Deaths:')

The top is a simple program to input user data, in this program it will input a string, and for the program to break right away, we use:

if n1.isdigit and n2.isdigit and n2.isdigit:
    n1 = float (n1)
    n2 = float (n2)
    n3 = float (n3)

That is, he is saying that if an input of n1, n2 and n3 is a number, convert it to me in float.

print (f'Total actually infected are: {n1 - (n2 + n3) :. 6} ')

If the user has not entered a letter, or an invalid value, he will jump directly to the calculation and display it on the screen!

   else:
    print ('It's not a number')
Otherwise, it will send this error message!